requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
bcrypt==4.1.2
openpyxl==3.1.2
SQLAlchemy==2.0.23
Werkzeug==3.0.1
python-dateutil==2.8.2
